+++ api/urls.py
@@ -8,6 +8,7 @@ urlpatterns = [
     url(r'^account/', include('api.accounts.urls')),
     url(r'^tournaments/', include('api.tournaments.urls')),
     url(r'^teams/', include('api.team.urls')),
+    url(r'^players/', include('api.players.urls')),
     url(r'^auth/', include('rest_framework.urls', namespace='rest_framework')),
     url('^', NotFoundView.raise_404)
 ]